Add interruption of ZipUnarchiver
The `Resource` used for the ZipEntry in ZipUnarchiver did not
allow for interruption because the `Async[F].blocking` call to
the underlying `ZipInputStream.closeEntry()` is not interruptable
and will block until the internal buffer in the ZipInputStream is
completely emptied.
Unfortunately signalling an interrupt to the `ZipInputStream` via
`Async[F].interruptible or `Async[F].interruptibleMany` doesn't
work either.

As a workaround we simply forgo closing the ZipEntry in case the
resource is released for any other reason than ExitCase.Succeeded